+++ CHANGELOG.md
@@ -2,13 +2,20 @@
 
 ## Version 1.1.0 - [TBD]
 
-- Features
+- Features:
   - Macro action bar positions are saved with set.
   - Macros are placed on action bars on set load.
-- Changes
+  - Confirmed working with Bartender4 and ElvUI.
+- Changes:
   - Macro set name character limit increased from **25** to **50**.
   - Checks to restrict addon usage during combat.
   - Checks for duplicate macro names when saving a set.
+  - Added help information for `/ms icons` to `/ms help`.
+- Commands:
+  - `/ms icons`: Toggles what the `#i` flag does at the end of macro names
+    - Set all macro icons to the default icon when saved if toggled to `on`.
+    - Set all macro icons to the currently displayed icon when saved if toggled `off`.
+    - Set to `off` by default.
 - Added simple testing framework to aid with debugging.
 - Substantial code refactoring.